requirements.txt
psutil
numpy
natsort
tqdm
scipy
pynrrd
nibabel
git+https://github.com/spezold/mvloader.git
vtk
SimpleITK
meshio
gmsh
scikit-image
scikit-fuzzy
#
absl-py
astunparse
cachetools
certifi
charset-normalizer
contourpy
cycler
fonttools
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
importlib-metadata
Keras
Keras-Applications
Keras-Preprocessing
kiwisolver
Markdown
MarkupSafe
matplotlib
nibabel
numpy
oauthlib
opt-einsum
packaging
Pillow
protobuf
pyasn1
pyasn1-modules
pyparsing
python-dateutil
PyYAML
requests
requests-oauthlib
rsa
scipy
six
tensorboard
tensorboard-plugin-wit
tensorflow-estimator
termcolor
urllib3
Werkzeug
wrapt
zipp